+++ lib/hyperclient/resource.rb
@@ -33,7 +33,7 @@ module Hyperclient
     # entry_point    - The EntryPoint object to inject the configutation.
     def initialize(representation, entry_point, response = nil)
       representation = representation ? representation.dup : {}
-      @_links       = LinkCollection.new(representation['_links'], entry_point)
+      @_links       = LinkCollection.new(representation['_links'], representation['_curies'], entry_point)
       @_embedded    = ResourceCollection.new(representation['_embedded'], entry_point)
       @_attributes  = Attributes.new(representation)
       @_entry_point = entry_point